Implemented MapSS class & code compiles. Need to make an episode player class to debug
Former-commit-id: 34da92a3ecb26a0172a07d8aad08073e880267a7

+++ project/Starship/waypoint.h
@@ -1,5 +1,6 @@
 #ifndef WAYPOINT_H
 #define WAYPOINT_H
+
 #include "vect2d.h"
 
 class Waypoint
@@ -10,7 +11,7 @@ class Waypoint
         double getRadius();
         int getColor();
         void setCentre(Vect2d u);
-        void setRadius(double r);
+        void setRadius(float r);
         void setColor(int color);
 
     private:
